pyproject.toml
[project]
name = "hmpps-sharepoint-discovery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity==1.25.2",
    "dockerfile-parse==2.0.1",
    "hmpps-sre-python-lib",
]

[tool.uv.sources]
hmpps-sre-python-lib = { git = "https://github.com/ministryofjustice/hmpps-sre-python-lib.git", rev = "feat/sharepoint_class" }
